pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["David Reifferscheidt <david.reifferscheidt@tum.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
google-generativeai = "^0.8.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
